requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
chompjs==1.4.0
curl_cffi==0.15.0
greenlet==3.4.0
idna==3.12
markdown-it-py==4.0.0
mdurl==0.1.2
playwright==1.58.0
pycparser==3.0
pyee==13.0.1
Pygments==2.20.0
python-dotenv==1.2.2
requests==2.33.1
rich==15.0.0
typing_extensions==4.15.0
urllib3==2.6.3
